add debian files that were accidentally omitted
This was re #961. An earlier commit intended to close that ticket accidentally
omitted these files. Later (#1454), we decided to remove all debian config
files from the upstream Tahoe-LAFS distribution, therefore this branch is no
longer suitable for merging to trunk. I'm creating this branch solely as a way
to give us and Debian packagers a common codebase to point to.