release sqlite db fds whenever LeaseDB is stopped
Brian correctly diagnosed this issue and suggested the fix in the Weekly Dev
Chat of 2013-07-09. This fix has been manually tested by me using fdleakfinder.
Without this patch, the 1819-cloud-merge branch uses up more than 1020 fds and
starts failing. With this patch the high water mark is 47 fds (because there is
a single big test that creates many LeaseDB instances).

However, this patch doesn't yet have a unit test, so I would advise not merging
this branch to trunk yet.