magic first line tells emacs to use utf8+bom
Add ".. -*- coding: utf-8-with-signature -*-" to the first line of each .rst
file in docs/. This tells emacs to treat the file contents as utf-8, and also
to prepend a so-called utf-8 "bom" marker at the beginning of the file. This
patch also prepends those markers to each of those files.