Rip out the "retry" feature of dirnodes.
This eliminates backoffer, first_time, and sometimes servermap arguments to
functions which no longer needs them. It collapses a couple of parts of the
call stack where one function now has almost nothing left to do but call
another function.

However, it does not attempt to refactor the code to have a simpler structure
in light of the fact that the code no longer does retrying.

There is an incomplete mockery-based test added. There is an extant test that
is broken by this patch.

This is a Work In Progress branch -- definitely not appropriate to get merged
into any other branch, especially not anybody's trunk.