fix one test, make a new test that is red
Fix the first test. It is green. I edited the code-under-test to write out the
wrong data and it made the test turn red.

Added a new test which is supposed to exercise #1670. It is red. Not sure if
that means the test is run or the code-under-test is still exhibiting #1670
even though I ripped out retrying. (The fact is that the code does
servermap-update more times than it ought to even *without* retrying, so it is
entirely possible that the current code-under-tets is still harboring an
instance of the #1670 bug and this test is demonstrating that.)