This patch is incomplete/wrong.
For one thing, it assumes that the version of foolscap it is using has
https://github.com/warner/foolscap/pull/20 applied, so that Tub.listenOn()
accepts unicodes instead of strs. To make it work with *real* foolscap it needs
to be changed to utf-8 encode the argument to listenOn().

For another thing, this causes unit tests to fail for a reason I don't exactly
understand, but it has to do with the swissnum for the introducer being a
unicode, and the remote foolscap side rejecting that during its schema
type-check since it requires incoming swissnums to be strs.

The part that I don't understand is where the outgoing swissnum is being sent,
because that, I guess, is where I want it to be encoded into utf-8…